+++ administrator/components/com_guidedtours/src/Helper/GuidedtoursHelper.php
@@ -51,4 +51,28 @@ class GuidedtoursHelper
 
         return $db->loadObject();
     }
+
+    public static function setStepLanguage(int $id, string $language): string
+    {
+        if (empty($id)) {
+            return false;
+        }
+
+        $db = Factory::getDbo();
+        $query = $db->getQuery(true);
+
+        $fields = array(
+            $db->quoteName('language') . ' = ' . $db->quote($language),
+        );
+
+        $conditions = array(
+            $db->quoteName('tour_id') . ' = ' . $db->quote($id)
+        );
+
+        $query->update($db->quoteName('#__guidedtour_steps'))->set($fields)->where($conditions);
+
+        $db->setQuery($query);
+
+        return $db->execute();
+    }
 }
