+++ administrator/components/com_guidedtours/src/Helper/StepHelper.php
@@ -51,29 +51,4 @@ class StepHelper
 
         return $db->loadResult();
     }
-
-    // Set step language.
-    public static function setStepLanguage(int $id, string $language): string
-    {
-        if (empty($id)) {
-            return false;
-        }
-
-        $db = Factory::getDbo();
-        $query = $db->getQuery(true);
-
-        $fields = array(
-            $db->quoteName('language') . ' = ' . $db->quote($language),
-        );
-
-        $conditions = array(
-            $db->quoteName('tour_id') . ' = ' . $db->quote($id)
-        );
-
-        $query->update($db->quoteName('#__guidedtour_steps'))->set($fields)->where($conditions);
-
-        $db->setQuery($query);
-
-        return $db->execute();
-    }
 }
